Refactor to include state setups for FInRem COS in one file
