+++ src/contractTest/java/uk/gov/hmcts/reform/managecase/api/controller/provider/NocCaseAssignmentProviderTests.java
@@ -17,7 +17,9 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import uk.gov.hmcts.reform.idam.client.models.UserDetails;
 import uk.gov.hmcts.reform.managecase.api.controller.CaseAssignmentController;
+import uk.gov.hmcts.reform.managecase.api.controller.NoticeOfChangeController;
 import uk.gov.hmcts.reform.managecase.client.datastore.CaseUserRole;
+import uk.gov.hmcts.reform.managecase.client.prd.FindOrganisationResponse;
 import uk.gov.hmcts.reform.managecase.config.MapperConfig;
 import uk.gov.hmcts.reform.managecase.domain.OrganisationPolicy;
 import uk.gov.hmcts.reform.managecase.repository.DataStoreRepository;
@@ -29,11 +31,13 @@ import uk.gov.hmcts.reform.managecase.util.JacksonUtils;
 import java.io.IOException;
 import java.util.List;
 
+import static java.util.Collections.emptyList;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.BDDMockito.given;
+import static org.mockito.Mockito.when;
 import static uk.gov.hmcts.reform.managecase.TestFixtures.CaseDetailsFixture.organisationPolicy;
 import static uk.gov.hmcts.reform.managecase.TestFixtures.ProfessionalUserFixture.user;
 import static uk.gov.hmcts.reform.managecase.TestFixtures.ProfessionalUserFixture.usersByOrganisation;
@@ -43,10 +47,10 @@ import static uk.gov.hmcts.reform.managecase.TestFixtures.ProfessionalUserFixtur
 @PactBroker(scheme = "${PACT_BROKER_SCHEME:http}",
     host = "${PACT_BROKER_URL:localhost}",
     port = "${PACT_BROKER_PORT:80}",
-    consumerVersionSelectors = {@VersionSelector(tag = "master")})
+    consumerVersionSelectors = {@VersionSelector(tag = "Dev")})
 @ContextConfiguration(classes = {ContractConfig.class, MapperConfig.class})
 @IgnoreNoPactsToVerify
-public class CasesAssignmentControllerProviderTest {
+public class NocCaseAssignmentProviderTests {
 
     private static final String ORG_POLICY_ROLE = "caseworker-probate";
     private static final String ORG_POLICY_ROLE2 = "caseworker-probate2";
@@ -63,6 +67,9 @@ public class CasesAssignmentControllerProviderTest {
     private static final String CASE_ROLE = "[CR1]";
     private static final String CASE_ROLE2 = "[CR2]";
 
+    private static final String TEST_APP_ORG_ID = "appOrgId";
+    private static final String TEST_APP_ORG_NAME = "appOrgName";
+
     @Autowired
     DataStoreRepository dataStoreRepository;
     @Autowired
@@ -77,6 +84,9 @@ public class CasesAssignmentControllerProviderTest {
     @Autowired
     CaseAssignmentController caseAssignmentController;
 
+    @Autowired
+    NoticeOfChangeController noticeOfChangeController;
+
 
     @TestTemplate
     @ExtendWith(PactVerificationInvocationContextProvider.class)
@@ -90,7 +100,7 @@ public class CasesAssignmentControllerProviderTest {
     void before(PactVerificationContext context) {
         MockMvcTestTarget testTarget = new MockMvcTestTarget();
         //System.getProperties().setProperty("pact.verifier.publishResults", "true");
-        testTarget.setControllers(caseAssignmentController);
+        testTarget.setControllers(caseAssignmentController, noticeOfChangeController);
         if (context != null) {
             context.setTarget(testTarget);
         }
@@ -137,5 +147,20 @@ public class CasesAssignmentControllerProviderTest {
                 new CaseUserRole(CASE_ID2, ASSIGNEE_ID2, CASE_ROLE),
                 new CaseUserRole(CASE_ID2, ASSIGNEE_ID2, CASE_ROLE2)));
 
+    }
+
+    @State({"A notice of change against case"})
+    public void toApplyNoticeOfChange() throws IOException {
+        given(dataStoreRepository.getCaseAssignments(any(), any()))
+            .willReturn(List.of(
+                new CaseUserRole(CASE_ID, ASSIGNEE_ID, CASE_ROLE),
+                new CaseUserRole(CASE_ID2, ASSIGNEE_ID2, CASE_ROLE2)
+            ));
+
+        given(prdRepository.findUsersByOrganisation(anyString()))
+            .willReturn(usersByOrganisation(user(ASSIGNEE_ID), user(ASSIGNEE_ID2)));
+
+        when(prdRepository.findOrganisationAddress(any()))
+            .thenReturn(new FindOrganisationResponse(emptyList(), TEST_APP_ORG_ID, TEST_APP_ORG_NAME));
     }
 }
